+++ src/twat_fs/upload_providers/litterbox.py
@@ -15,7 +15,7 @@ from typing import Any, BinaryIO, ClassVar, cast
 import aiohttp
 from loguru import logger
 
-from twat_fs.upload_providers.protocols import ProviderClient, Provider, ProviderHelp
+from twat_fs.upload_providers.protocols import ProviderClient, ProviderHelp
 from twat_fs.upload_providers.simple import BaseProvider
 from twat_fs.upload_providers.types import ExpirationTime, UploadResult
 from twat_fs.upload_providers.core import (
@@ -28,7 +28,6 @@ from twat_fs.upload_providers.utils import (
     create_provider_help,
     log_upload_attempt,
     handle_http_response,
-    standard_upload_wrapper,
 )
 
 LITTERBOX_API_URL = "https://litterbox.catbox.moe/resources/internals/api.php"